+++ test/tests/runner.cljs
@@ -3,7 +3,7 @@
             [cljsjs.react]
             [cljsjs.react.dom]
             [re-view.core-test]
-            [re-view.subscriptions-test]))
+            [re-view.state-test]))
 
 (doo-tests 're-view.core-test
-           're-view.subscriptions-test)
+           're-view.state-test)